docs: Standrard text file for documentation.
Testing/practicing commit messages with Atom text editor.

More detailed explanatory text, if necessary. Wrap it to about 72
characters or so. In some contexts, the first line is treated as the
subject of the commit and the rest of the text as the body. The
blank line separating the summary from the body is critical (unless
you omit the body entirely); various tools like `log`, `shortlog`
and `rebase` can get confused if you run the two together.

- bullet point
- bullet point

No issue tracker in place at this time:

Resolves: #999
See also: #666